Forwarder supports "value"
forwarder now has a "value" parameter, which is the amount of ether to
send with the transaction.
The `execute()` function is now payable, to be able to receive
this value.
Note that it is possible to pass the value into the forwarder in a call
just before execute (still, under the same transaction)

The `execute()` would fail the request if it is given a value
parameter, but doesn't have enough value to fulfill it.